:hammer: refactor hex bitmaps into a separate file
