+++ src/DiffViewerProvider.ts
@@ -1,5 +1,6 @@
 import { Diff2HtmlConfig, parse } from "diff2html";
 import { LineMatchingType, OutputFormatType } from "diff2html/lib/types";
+import { parseHexBitmap, bitmapToHex } from "./hex-bitmap";
 import * as vscode from "vscode";
 
 export class DiffViewerProvider implements vscode.CustomTextEditorProvider {
@@ -181,46 +182,3 @@ function updateViewedFiles(content: string, update: FileViewedMessage, eol: stri
 }
 
 
-const HEX: {[s: string]: [boolean, boolean, boolean, boolean]} = {
-  '0': [false, false, false, false],
-  '1': [false, false, false, true],
-  '2': [false, false, true, false],
-  '3': [false, false, true, true],
-  '4': [false, true, false, false],
-  '5': [false, true, false, true],
-  '6': [false, true, true, false],
-  '7': [false, true, true, true],
-  '8': [true, false, false, false],
-  '9': [true, false, false, true],
-  'a': [true, false, true, false],
-  'b': [true, false, true, true],
-  'c': [true, true, false, false],
-  'd': [true, true, false, true],
-  'e': [true, true, true, false],
-  'f': [true, true, true, true],
-}
-
-const TO_HEX = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'];
-
-function parseHexBitmap(s: string): boolean[] {
-  const retval = [];
-  for (const char of s) {
-    retval.push(...HEX[char]);
-  }
-  return retval;
-}
-
-function bitmapToHex(bitmap: boolean[]): string {
-  const retval = [];
-
-  // go 4 bits (a nibble) at a time
-  for (let i = 0; i < bitmap.length; i += 4) {
-    const number = (bitmap[i] ? 8 : 0) +
-      (bitmap[i+1] ? 4 : 0) +
-      (bitmap[i+2] ? 2 : 0) +
-      (bitmap[i+3] ? 1 : 0);
-    retval.push(TO_HEX[number]);
-  }
-
-  return retval.join('');
-}
